refactor(ai): phase 11 slice A — split near-cap planning libs and ratchet to 400 (#4239)
Concern-split the eight baseline near-cap colonizethis_ai planning modules
and lower repo.ai_source_file_size from 450 to 400 physical lines so additive
work retains headroom without behavior changes.

Refs #4239
