+++ SPEC/program/ai-source-file-size.md
@@ -1,18 +1,18 @@
 # AI package — lib/src physical line size ratchet (repo lint)
 
 **SPEC/program** — repository lint gate that keeps every hand-written Dart
-file under `packages/colonizethis_ai/lib/src/` at or below **450 physical
+file under `packages/colonizethis_ai/lib/src/` at or below **400 physical
 lines**, with an optional shrink-only grandfather allowlist.
 
 ## Motivation
 
 Phase 9 (#4079) concern-split planning modules that had grown past 500
 physical lines. Phase 10 Slice B (#4104) concern-split the five near-cap
-modules (declare-war suppression, research funding/decision, recruitment
-ledger, treasury regiment bids, economy cargo) and ratcheted the ceiling
-to **450** so those siblings cannot silently re-merge toward the prior
-500 band. This gate mirrors the save/data shrink-only allowlist pattern.
-AI is not an extracted logic-domain package, so it is not listed under
+modules and ratcheted the ceiling to **450**. Phase 11 Slice A (#4239)
+concern-split the eight remaining near-cap planning modules and ratcheted
+the ceiling to **400** so additive feature work retains headroom. This gate
+mirrors the save/data shrink-only allowlist pattern. AI is not an extracted
+logic-domain package, so it is not listed under
 `repo.domain_package_source_file_size` (still 500 elsewhere).
 
 ## Source of truth
@@ -38,14 +38,15 @@ temporarily exceed the ceiling. The allowlist is **shrink-only**:
 - Path exists and is now ≤ ceiling → fail (stale entry; remove from allowlist).
 - Path exists and is still over ceiling → skip size violation for that path.
 
-After #4079 Slice C / #4104 Slice B the production allowlist is empty.
+After #4079 Slice C / #4104 Slice B / #4239 Slice A the production allowlist
+is empty.
 
 ## Acceptance criteria
 
 - Given every hand-written file under `packages/colonizethis_ai/lib/src/**` is
-  at or below 450 physical lines and the grandfather allowlist is empty, when
+  at or below 400 physical lines and the grandfather allowlist is empty, when
   the System runs `runCheckAiSourceFileSize`, then the System exits `0`.
-- Given a hand-written AI `lib/src` file with more than 450 physical lines and
+- Given a hand-written AI `lib/src` file with more than 400 physical lines and
   an empty grandfather allowlist, when the System runs `runCheckAiSourceFileSize`,
   then the System exits non-zero and names that file.
 - Given an over-cap AI `lib/src` file listed on the grandfather allowlist, when
@@ -57,6 +58,6 @@ After #4079 Slice C / #4104 Slice B the production allowlist is empty.
 - Given a grandfather allowlist entry that names a file now at or below the
   ceiling, when the System runs `runCheckAiSourceFileSize`, then the System
   exits non-zero and reports that the entry must be removed from the allowlist.
-- Given a generated AI `lib/src` file (for example `*.g.dart`) over 450
+- Given a generated AI `lib/src` file (for example `*.g.dart`) over 400
   physical lines, when the System runs `runCheckAiSourceFileSize`, then the
   System does not fail because of that generated file.
